// src/models/services.ts
export interface Services {
  city?: string;

  servicelink1?: string;
  place1?: string;
  servicePhone1?: string;

  servicelink2?: string;
  place2?: string;
  servicePhone2?: string;

  servicelink3?: string;
  place3?: string;
  servicePhone3?: string;

  servicelink4?: string;
  place4?: string;
  servicePhone4?: string;

  servicelink5?: string;
  place5?: string;
  servicePhone5?: string;

  titleB1?: string;
  placeB1?: string;
  titleB2?: string;
  placeB2?: string;
  titleB3?: string;
  placeB3?: string;
  titleB4?: string;
  placeB4?: string;
  titleB5?: string;
  placeB5?: string;
  titleB6?: string;
  placeB6?: string;
  titleB7?: string;
  placeB7?: string;
  titleB8?: string;
  placeB8?: string;
  titleB9?: string;
  placeB9?: string;
  titleB10?: string;
  placeB10?: string;
  titleB11?: string;
  placeB11?: string;
  titleB12?: string;
  placeB12?: string;
  titleB13?: string;
  placeB13?: string;
  titleB14?: string;
  placeB14?: string;

  placeR1?: string;
  titleR1?: string;
  placeR2?: string;
  titleR2?: string;
  placeR3?: string;
  titleR3?: string;
  placeR4?: string;
  titleR4?: string;
  placeR5?: string;
  titleR5?: string;
  placeR6?: string;
  titleR6?: string;
}